backend/pyproject.toml
[tool.poetry]
name = "autometabuilder"
version = "0.1.0"
description = "AutoMetabuilder"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "autometabuilder", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
PyGithub = "^2.1.1"
tenacity = "^9.1.2"
flask = "^2.3.3"
slack-sdk = "^3.39.0"
discord-py = "^2.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autometabuilder = "autometabuilder.main:main"

[dependency-groups]
dev = [
    "playwright (>=1.57.0,<2.0.0)",
    "pytest-playwright (>=0.7.2,<0.8.0)"
]
